# crates/iso-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "iso admin API",
    "description": "Create Firecracker microVMs, set their egress policy, and reach the agent inside them. Served on the daemon's unix socket (root only) and on its TCP listener, where a client certificate issued by the host's admin CA (`isoctl admin issue-client`) is required.",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/vms": {
      "get": {
        "tags": [
          "vms"
        ],
        "operationId": "list",
        "responses": {
          "200": {
            "description": "Every VM record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vm"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "vms"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created and booted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedVm"
                }
              }
            }
          },
          "400": {
            "description": "A malformed id or rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Unknown template",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The id exists, no free slot or port, or the pool is over its watermark",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/vms/{id}": {
      "get": {
        "tags": [
          "vms"
        ],
        "operationId": "get_one",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "VM id, as a hyphenated UUID or 32 hex digits",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The VM",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vm"
                }
              }
            }
          },
          "404": {
            "description": "Unknown VM",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "vms"
        ],
        "operationId": "destroy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "VM id, as a hyphenated UUID or 32 hex digits",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Destroyed with its storage and placement, whatever its lifecycle"
          },
          "404": {
            "description": "Unknown VM",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Not valid in the VM's current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/vms/{id}/policy": {
      "patch": {
        "tags": [
          "vms"
        ],
        "operationId": "set_policy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "VM id, as a hyphenated UUID or 32 hex digits",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Applied to new connections; connections at the old policy generation are closed by the proxy edge"
          },
          "400": {
            "description": "A rule does not parse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Unknown VM",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateVmRequest": {
        "type": "object",
        "description": "Create and boot a VM. Only `template` is required.",
        "required": [
          "template"
        ],
        "properties": {
          "allow": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Domains routed through the egress proxy."
          },
          "egress": {
            "type": "string",
            "description": "`proxy` or `deny` (default). Unrecognized values mean `deny`. There is\nno direct mode: how much a VM may reach is its rules' business\n(`allow https://*/**` opens everything, through the proxy)."
          },
          "id": {
            "type": "string",
            "description": "Create under this id (hyphenated UUID or 32 hex digits) instead of a\ngenerated one. A fleet service records the id before it calls, so a\nretry after a lost reply is a `409`, never a second VM.",
            "nullable": true
          },
          "ingress": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortForward"
            },
            "description": "Ports to forward into the VM; host ports are allocated."
          },
          "labels": {
            "type": "object",
            "description": "Free-form labels; `name` is surfaced to the guest's metadata.",
            "additionalProperties": {
              "type": "string"
            }
          },
          "lifecycle": {
            "type": "string",
            "description": "`ephemeral` (default: deleted on stop) or `durable` (rootfs kept)."
          },
          "mem_mib": {
            "type": "integer",
            "format": "int32",
            "description": "Override the template's memory (forces a cold boot).",
            "nullable": true,
            "minimum": 0
          },
          "principal": {
            "type": "string",
            "description": "Principal whose credentials the egress proxy injects.",
            "nullable": true
          },
          "restart": {
            "type": "string",
            "description": "`never` (default), `on_failure` or `always`."
          },
          "rules": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "URI-level rules, e.g. `allow https://api.github.com/**` or\n`deny https://api.github.com/user/keys`. Deny wins; nothing matching\nis denied. `allow` hosts are sugar for `allow https://host/**` plus\n`allow wss://host/**`. A rule that does not parse is a 400."
          },
          "signed": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SignedPolicyDto"
              }
            ],
            "nullable": true
          },
          "template": {
            "type": "string",
            "description": "A registered template name."
          },
          "vcpus": {
            "type": "integer",
            "format": "int32",
            "description": "Override the template's vCPU count (forces a cold boot).",
            "nullable": true,
            "minimum": 0
          }
        }
      },
      "CreatedVm": {
        "type": "object",
        "description": "The id of a newly created VM.",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The shape of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong."
          }
        }
      },
...
      "PolicyRequest": {
        "type": "object",
        "description": "Change a running VM's egress policy. Omitted fields are left unchanged.",
        "properties": {
          "allow": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "egress": {
            "type": "string",
            "description": "`proxy` or `deny`; anything else leaves the mode unchanged.",
            "nullable": true
          },
          "principal": {
            "type": "string",
            "nullable": true
          },
          "rules": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Replace the URI-level rules. A rule that does not parse is a 400.",
            "nullable": true
          },
          "signed": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SignedPolicyDto"
              }
            ],
            "nullable": true
          }
        }
      },
      "PortForward": {
        "type": "object",
        "description": "An ingress port forward from a host port to a port inside the VM.",
        "required": [
          "vm_port",
          "proto"
        ],
        "properties": {
          "host_port": {
            "type": "integer",
            "format": "int32",
            "description": "Allocated by the control plane; ignored on create, reported on read.",
            "minimum": 0
          },
          "proto": {
            "type": "string",
            "description": "`tcp` or `udp`."
          },
          "vm_port": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "SignedPolicyDto": {
        "type": "object",
        "description": "A policy as a fleet signed it: base64 claims and an ed25519 signature\nover those bytes. Opaque here.",
        "required": [
          "claims",
          "sig"
        ],
        "properties": {
          "claims": {
            "type": "string"
          },
          "sig": {
            "type": "string"
          }
        }
      },
...
      "Vm": {
        "type": "object",
        "description": "A VM as the control plane sees it.",
        "required": [
          "id",
          "template",
          "state",
          "egress",
          "lifecycle",
          "restart",
          "labels",
          "ingress",
          "allow",
          "rules",
          "policy_gen"
        ],
        "properties": {
          "allow": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "egress": {
            "type": "string",
            "description": "`proxy` or `deny`."
          },
          "id": {
            "type": "string"
          },
          "ingress": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortForward"
            }
          },
          "labels": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "lifecycle": {
            "type": "string",
            "description": "`ephemeral` or `durable`."
          },
          "policy_gen": {
            "type": "integer",
            "format": "int64",
            "description": "Bumped on every policy change.",
            "minimum": 0
          },
          "principal": {
            "type": "string",
            "nullable": true
          },
          "restart": {
            "type": "string",
            "description": "`never`, `on_failure` or `always`."
          },
          "rootfs_device": {
            "type": "string",
            "nullable": true
          },
          "rules": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "URI-level rules on top of `allow`, in `iso-policy` syntax."
          },
          "signed": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SignedPolicyDto"
              }
            ],
            "nullable": true
          },
          "slot": {
            "type": "integer",
            "format": "int32",
            "description": "Placement slot; absent while a durable VM is stopped.",
            "nullable": true,
            "minimum": 0
          },
          "state": {
            "type": "string",
            "description": "`creating`, `running`, `suspended`, `stopped` or `failed`."
          },
          "tap": {
            "type": "string",
            "nullable": true
          },
          "template": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}